+++ backend/agent/api.py
@@ -596,7 +596,23 @@ async def get_thread_agent(thread_id: str, user_id: str = Depends(get_current_us
             except Exception as e:
                 logger.warning(f"Failed to get version data for agent {effective_agent_id}: {e}")
         
-        # Extract configuration using the unified config approach
+        version_data = None
+        if current_version:
+            version_data = {
+                'version_id': current_version.version_id,
+                'agent_id': current_version.agent_id,
+                'version_number': current_version.version_number,
+                'version_name': current_version.version_name,
+                'system_prompt': current_version.system_prompt,
+                'configured_mcps': current_version.configured_mcps,
+                'custom_mcps': current_version.custom_mcps,
+                'agentpress_tools': current_version.agentpress_tools,
+                'is_active': current_version.is_active,
+                'created_at': current_version.created_at,
+                'updated_at': current_version.updated_at,
+                'created_by': current_version.created_by
+            }
+        
         from agent.config_helper import extract_agent_config
         agent_config = extract_agent_config(agent_data, version_data)
         
@@ -2125,15 +2141,15 @@ async def get_pipedream_tools_for_agent(
     logger.info(f"Getting tools for agent {agent_id}, profile {profile_id}, user {user_id}, version {version}")
 
     try:
-        from pipedream.facade import PipedreamManager
-        pipedream_manager = PipedreamManager()
+        from pipedream import profile_service, mcp_service
+        from uuid import UUID
 
-        profile = await pipedream_manager.get_profile(user_id, profile_id)
+        profile = await profile_service.get_profile(UUID(user_id), UUID(profile_id))
         
         if not profile:
             logger.error(f"Profile {profile_id} not found for user {user_id}")
             try:
-                all_profiles = await pipedream_manager.get_profiles(user_id)
+                all_profiles = await profile_service.get_profiles(UUID(user_id))
                 pipedream_profiles = [p for p in all_profiles if 'pipedream' in p.mcp_qualified_name]
                 logger.info(f"User {user_id} has {len(pipedream_profiles)} pipedream profiles: {[p.profile_id for p in pipedream_profiles]}")
             except Exception as debug_e:
@@ -2144,29 +2160,62 @@ async def get_pipedream_tools_for_agent(
         if not profile.is_connected:
             raise HTTPException(status_code=400, detail="Profile is not connected")
 
-        if version:
-            enabled_tools = await pipedream_manager.get_enabled_tools_for_agent_profile_version(
-                agent_id=agent_id,
-                profile_id=profile_id,
-                user_id=user_id,
-                version_id=version
-            )
-            logger.info(f"[VERSION {version}] Retrieved {len(enabled_tools)} enabled tools: {enabled_tools}")
-        else:
-            enabled_tools = await pipedream_manager.get_enabled_tools_for_agent_profile(
-                agent_id=agent_id,
-                profile_id=profile_id,
-                user_id=user_id
-            )
-            logger.info(f"[CURRENT VERSION] Retrieved {len(enabled_tools)} enabled tools: {enabled_tools}")
+        enabled_tools = []
+        try:
+            client = await db.client
+            agent_row = await client.table('agents')\
+                .select('current_version_id')\
+                .eq('agent_id', agent_id)\
+                .eq('account_id', user_id)\
+                .maybe_single()\
+                .execute()
+            
+            if agent_row.data and agent_row.data.get('current_version_id'):
+                if version:
+                    version_result = await client.table('agent_versions')\
+                        .select('config')\
+                        .eq('version_id', version)\
+                        .maybe_single()\
+                        .execute()
+                else:
+                    version_result = await client.table('agent_versions')\
+                        .select('config')\
+                        .eq('version_id', agent_row.data['current_version_id'])\
+                        .maybe_single()\
+                        .execute()
+                
+                if version_result.data and version_result.data.get('config'):
+                    agent_config = version_result.data['config']
+                    tools = agent_config.get('tools', {})
+                    custom_mcps = tools.get('custom_mcp', []) or []
+                    
+                    for mcp in custom_mcps:
+                        mcp_profile_id = mcp.get('config', {}).get('profile_id')
+                        if mcp_profile_id == profile_id:
+                            enabled_tools = mcp.get('enabledTools', mcp.get('enabled_tools', []))
+                            logger.info(f"Found enabled tools for profile {profile_id}: {enabled_tools}")
+                            break
+                    
+                    if not enabled_tools:
+                        logger.info(f"No enabled tools found for profile {profile_id} in agent {agent_id}")
+            
+        except Exception as e:
+            logger.error(f"Error retrieving enabled tools for profile {profile_id}: {str(e)}")
+        
+        logger.info(f"Using {len(enabled_tools)} enabled tools for profile {profile_id}: {enabled_tools}")
         
         try:
-            servers = await pipedream_manager.discover_mcp_servers(
-                external_user_id=profile.external_user_id,
-                app_slug=profile.app_slug
-            )
+            from pipedream.mcp_service import ExternalUserId, AppSlug
+            external_user_id = ExternalUserId(profile.external_user_id.value)
+            app_slug_obj = AppSlug(profile.app_slug.value)
             
-            server = next((s for s in servers if s.app_slug == profile.app_slug), None)
+            logger.info(f"Discovering servers for user {external_user_id.value} and app {app_slug_obj.value}")
+            servers = await mcp_service.discover_servers_for_user(external_user_id, app_slug_obj)
+            logger.info(f"Found {len(servers)} servers: {[s.app_slug.value for s in servers]}")
+            
+            server = servers[0] if servers else None
+            logger.info(f"Selected server: {server.app_slug.value if server else 'None'} with {len(server.available_tools) if server else 0} tools")
+            
             if not server:
                 return {
                     'profile_id': profile_id,
@@ -2254,16 +2303,36 @@ async def update_pipedream_tools_for_agent(
 
         enabled_tools = request.get('enabled_tools', [])
         
-        from pipedream.facade import PipedreamManager
-        pipedream_manager = PipedreamManager()
+        updated = False
+        for mcp in custom_mcps:
+            mcp_profile_id = mcp.get('config', {}).get('profile_id')
+            if mcp_profile_id == profile_id:
+                mcp['enabledTools'] = enabled_tools
+                mcp['enabled_tools'] = enabled_tools
+                updated = True
+                logger.info(f"Updated enabled tools for profile {profile_id}: {enabled_tools}")
+                break
         
-        result = await pipedream_manager.update_agent_profile_tools(
-            agent_id=agent_id,
-            profile_id=profile_id,
-            user_id=user_id,
-            enabled_tools=enabled_tools
-        )
-        logger.info(f"Successfully updated Pipedream tools for agent {agent_id}, created version {result.get('version_name')}")
+        if not updated:
+            logger.warning(f"Profile {profile_id} not found in agent {agent_id} custom_mcps configuration")
+            
+        if updated:
+            agent_config['tools']['custom_mcp'] = custom_mcps
+            
+            await client.table('agent_versions')\
+                .update({'config': agent_config})\
+                .eq('version_id', agent_row.data['current_version_id'])\
+                .execute()
+            
+            logger.info(f"Successfully updated agent configuration for {agent_id}")
+        
+        result = {
+            'success': updated,
+            'enabled_tools': enabled_tools,
+            'total_tools': len(enabled_tools),
+            'profile_id': profile_id
+        }
+        logger.info(f"Successfully updated Pipedream tools for agent {agent_id}, profile {profile_id}")
         return result
         
     except ValueError as e:
@@ -2288,8 +2357,7 @@ async def get_custom_mcp_tools_for_agent(
             raise HTTPException(status_code=404, detail="Agent not found")
         
         agent = agent_result.data[0]
-        
-        # Get current version config
+ 
         agent_config = {}
         if agent.get('current_version_id'):
             version_result = await client.table('agent_versions')\
