+++ backend/agent/tools/utils/custom_mcp_handler.py
@@ -61,12 +61,11 @@ class CustomMCPHandler:
         
         try:
             import os
-            from pipedream.facade import PipedreamManager
+            from pipedream import connection_service
             from mcp import ClientSession
             from mcp.client.streamable_http import streamablehttp_client
             
-            pipedream_manager = PipedreamManager()
-            http_client = pipedream_manager._http_client
+            http_client = connection_service._http_client
             
             access_token = await http_client._ensure_access_token()
