+++ backend/pipedream/__init__.py
@@ -1,29 +1,28 @@
 from utils.logger import logger
 from services.supabase import DBConnection
 
-from .facade import PipedreamManager
-from .domain.entities import (
-    Profile,
-    Connection,
-    App,
-    MCPServer,
-    MCPTool,
-    ConnectionStatus,
-    AuthType
-)
+from .profile_service import ProfileService, Profile
+from .connection_service import ConnectionService, Connection, AuthType
+from .app_service import AppService, App
+from .mcp_service import MCPService, MCPServer, MCPTool, ConnectionStatus
+from .connection_token_service import ConnectionTokenService
 
 db = DBConnection()
 
-pipedream_manager = PipedreamManager(
-    db=db,
-    logger=logger
-)
+profile_service = ProfileService(db=db, logger=logger)
+connection_service = ConnectionService(logger=logger)
+app_service = AppService(logger=logger)
+mcp_service = MCPService(logger=logger)
+connection_token_service = ConnectionTokenService(logger=logger)
 
 from . import api
-api.pipedream_manager = pipedream_manager
+api.profile_service = profile_service
+api.connection_service = connection_service
+api.app_service = app_service
+api.mcp_service = mcp_service
+api.connection_token_service = connection_token_service
 
 __all__ = [
-    'PipedreamManager',
     'Profile',
     'Connection', 
     'App',
@@ -31,6 +30,10 @@ __all__ = [
     'MCPTool',
     'ConnectionStatus',
     'AuthType',
-    'pipedream_manager',
+    'profile_service',
+    'connection_service',
+    'app_service',
+    'mcp_service',
+    'connection_token_service',
     'api'
 ]
