--- backend/pipedream/repositories/__init__.py
@@ -1,11 +1,0 @@
-from .profile_repository import SupabaseProfileRepository
-from .connection_repository import PipedreamConnectionRepository
-from .app_repository import PipedreamAppRepository
-from .mcp_server_repository import PipedreamMCPServerRepository
-
-__all__ = [
-    "SupabaseProfileRepository",
-    "PipedreamConnectionRepository",
-    "PipedreamAppRepository", 
-    "PipedreamMCPServerRepository"
-]
