--- backend/pipedream/services/__init__.py
@@ -1,21 +1,0 @@
-from .profile_service import ProfileService
-from .connection_service import ConnectionService
-from .app_service import AppService
-from .mcp_service import MCPService
-from .external_user_id_service import ExternalUserIdService
-from .mcp_qualified_name_service import MCPQualifiedNameService
-from .connection_token_service import ConnectionTokenService
-from .profile_configuration_service import ProfileConfigurationService
-from .connection_status_service import ConnectionStatusService
-
-__all__ = [
-    "ProfileService",
-    "ConnectionService", 
-    "AppService",
-    "MCPService",
-    "ExternalUserIdService",
-    "MCPQualifiedNameService",
-    "ConnectionTokenService",
-    "ProfileConfigurationService",
-    "ConnectionStatusService"
-]
